template/pyproject.toml
[project]
name = "mcp-boilerplate"
version = "0.1.0"
description = "MCP Server boilerplate — OAuth2 (Azure Entra) + API key auth"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "mcp[cli]==1.26.0",
    "psutil>=5.9.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
]
